Add Scala Native support to http4s module
