+++ http4s-backend/src/test/scalanative/sttp/client4/http4s/Http4sNativeHttpTest.scala
@@ -1,0 +1,16 @@
+package sttp.client4.http4s
+
+import cats.effect.IO
+import sttp.client4.Backend
+import sttp.client4.impl.cats.{CatsRetryTest, CatsTestBase}
+import sttp.client4.testing.HttpTest
+
+class Http4sNativeHttpTest extends HttpTest[IO] with CatsRetryTest with CatsTestBase {
+
+  override val backend: Backend[IO] =
+    Http4sBackend.usingDefaultEmberClientBuilder[IO]().allocated.unsafeRunSync()._1
+
+  override protected def supportsRequestTimeout = false
+  override protected def supportsCustomMultipartContentType = false
+  override protected def supportsResponseAsInputStream = false
+}
